Ensure data is always BINARY-encoded in BSON::Binary objects.
This unbreaks serialization of BSON::Binary objects constructed
from frozen strings, since encoding is no longer changed during
serialization (which also used to mutate the constructor argument,
long after it was used in the constructor).

This change can break applications that used to read the data
of BSON::Binary expecting the data to be in whatever text encoding
the Binary was constructed with, but this usage is incorrect -
the applications should assume that when a Binary is constructed,
the data string given to it is relabeled as a byte string.